Abstract push server API with PushClient
